adding authetication middleware

+++ backend/controller/goal-controller.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User =require('../models/userModel')
 
 
 // @desc gets goals
 // @route GET/api/goals
 // @access Private
-const getGoals = asyncHandler( async (req,res) =>{
-    const goals = await Goal.find();
+const getGoals = asyncHandler( async (req,res) =>{    
+    const goals = await Goal.find({user:req.user.id});   
     res.status(200).json(goals)
 })
 
@@ -15,8 +16,17 @@ const getGoals = asyncHandler( async (req,res) =>{
 // @route DELETE/api/goals/:id
 // @access Private
 const deleteGoal = asyncHandler( async (req,res) =>{
-    const goal = await Goal.findByIdAndRemove(req.params.id)
-    res.status(200).json(goal)
+    const goal = await Goal.findById(req.params.id);
+    if(!goal){
+        res.status(401)
+        throw new Error(`Goal with ${req.params.id} doesn't exist`)
+    }
+    if(goal.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error("can't delete un owned goal")
+    }
+    const deletedGoal = await Goal.findByIdAndRemove(req.params.id)
+    res.status(200).json(deletedGoal)
 })
 
 
@@ -30,6 +40,7 @@ const addGoal = asyncHandler( async (req,res)=>{
     }
     const goal = await Goal.create({
         text:req.body.text,
+        user:req.user.id,
     })
     res.status(200).json(goal)
     
@@ -44,11 +55,17 @@ const updateGoal = asyncHandler( async (req,res)=>{
         res.status(400)
         throw new Error('Please enter a text field')
     }
+    
     const goal = await Goal.findById(req.params.id);
     if(!goal){
-        res.status(400)
-        throw new Error("goal doesn't exist")
+        res.status(401)
+        throw new Error(`Goal with ${req.params.id} doesn't exist`)
     }
+    if(goal.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error("can't unpdate un owned goal")
+    }
+
     const updatedgoal= await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updatedgoal)
 })
